Fixed all sorts of issues after a complicated rebase
