+++ src/lib/urls.js
@@ -88,7 +88,7 @@ module.exports = {
     },
     advisers: {
       index: url('/companies', '/:companyId/advisers'),
-      confirm: url('/companies', '/:companyId/advisers/confirm'),
+      confirm: url('/companies', '/:companyId/advisers/add'),
       replace: url('/companies', '/:companyId/advisers/replace'),
     },
     index: url('/companies'),
