+++ test/functional/cypress/specs/companies/lead-advisers/company-page-spec.js
@@ -1,5 +1,8 @@
+import { assertBreadcrumbs } from '../../../support/assertions'
+
 const fixtures = require('../../../fixtures')
 const selectors = require('../../../../../selectors')
+const { companies, dashboard } = require('../../../../../../src/lib/urls')
 
 describe('Lead advisers', () => {
   context('when viewing a non One List tier company', () => {
